+++ packages/analytics-js-common/src/constants/integrations/SpotifyPixel/constants.ts
@@ -1,0 +1,20 @@
+const DIR_NAME = 'SpotifyPixel';
+const NAME = 'SPOTIFYPIXEL';
+const DISPLAY_NAME = 'Spotify Pixel';
+
+const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
+const CNameMapping = {
+  [NAME]: NAME,
+  'Spotify Pixel': NAME,
+  'spotify pixel': NAME,
+  SPOTIFY_PIXEL: NAME,
+};
+
+export {
+  NAME,
+  CNameMapping,
+  DISPLAY_NAME_TO_DIR_NAME_MAP,
+  DISPLAY_NAME,
+  DIR_NAME,
+};
+