+++ packages/analytics-js/.size-limit.js
@@ -13,13 +13,13 @@ module.exports = [
     name: 'Core CJS - NPM',
     path: 'dist/npm/modern/cjs/index.js',
     gzip: true,
-    limit: '25.5 KiB',
+    limit: '26 KiB',
   },
   {
     name: 'Core - NPM',
     path: 'dist/npm/modern/umd/index.js',
     gzip: true,
-    limit: '25.5 KiB',
+    limit: '26 KiB',
   },
   {
     name: 'Core Legacy - CDN',