+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.48.0](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.47.0...rudder-sdk-js@2.48.0) (2024-01-19)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `2.47.0`
+
+### Features
+
+* **analytics-v1.1:** add lock integrations version as query param in source config url ([#1584](https://github.com/rudderlabs/rudder-sdk-js/issues/1584)) ([4163514](https://github.com/rudderlabs/rudder-sdk-js/commit/4163514b968fb33945e8b75c1c4f9a10dd4677fc))
+
 ## [2.47.0](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.46.0...rudder-sdk-js@2.47.0) (2024-01-08)
 
 ### Dependency Updates